deduplicate test suite
