--- apps/assessments/roar-levante-tasks/src/tests/__fixtures__/inference_lookup_v1.csv
@@ -1,10 +1,0 @@
-ageMonths,thetaEstimate,roarScore,standardScore,percentile
-72,-1.0,415,84,14
-72,0.0,475,98,45
-72,1.0,535,112,79
-108,-1.0,425,86,18
-108,0.0,485,99,47
-108,1.0,545,112,79
-216,-1.0,435,88,21
-216,0.0,495,100,50
-216,1.0,555,113,81
