--- apps/assessments/roar-levante-tasks/src/tests/__fixtures__/trog_lookup_v1.csv
@@ -1,10 +1,0 @@
-ageMonths,thetaEstimate,roarScore,standardScore,percentile
-72,-1.0,418,84,14
-72,0.0,478,99,47
-72,1.0,538,113,81
-108,-1.0,430,87,19
-108,0.0,490,100,50
-108,1.0,550,113,81
-216,-1.0,440,89,23
-216,0.0,500,100,50
-216,1.0,560,113,81
